services/user-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
psycopg2-binary
bcrypt